Specify a last modification date in the copyright header
You can't copyright something forever, it doesn't work that way. Ask
your local lawyer.

Abbreviated the "current year" variable name so that it has the same
number of characters as the year, so that it's easy to make the box
border line up.

Also, I fixed the chdir(), which was assuming some obsolete file layout.