Update license headers
Ran license.php with the changes made in the previous commit.

It's not possible to copyright something forever. Or at least, it is
not possible to copyright something forever just by writing "this is
copyright forever", you have to first get your legislature to change
the law, like what the UK parliament did for Peter Pan.

The year is there to advise potential infringers that the work was
changed at some particular date, and thus that the copyright has not
expired. If you don't put the year on there, then you don't get that
protective effect.