Predict that the result of SetM is the same as its input
If we can't prove at compile time that a set operation won't
fail, we don't know the exact type of SetM's stack output. Since these
failures should be extremely rare, we now predict that the set
operation won't fail and side exit if it did. This converts a bunch of
DecRefs to DecRefNZ, and when we get rid of all the SpillStacks for
the helper calls this will allow us to optimize away those
IncRef/DecRefNZ pairs (while pushing the IncRef to the exit trace).