Fix fb_intercept for functions with extra arguments
If you pass extra arguments to a function, and it doesn't
use any of the func_get_args type functions, by default we drop
them.

Now that fb_intercept runs after the prolog, rather than before
this means it doesnt see them. Disable the extra args optimization
when JitEnableRenameFunction is in effect.