Hoist CheckType instructions above generic IncRef/LdMem @override-unit-failures
Because we normally find out about a type prediction at hhbc
bounderies, sometimes we generate type checks after we've already done
some generic operations on Cells.  This optimizes a specific pattern
of this that occurs in vector translations.  The code is set up as if
there is more than one since I have another case partially
implemented, but it currently hits issues in both linearscan and
codegen that need to be addressed first, so I split the diff for now.