Reserve r11 for code-gen's scratch fallback
This diff reserves r11 for the GP scratch register used by the
CodeGenerator (in case no better register is available). rScratch
(r10) is now reserved for the assembler, and was renamed to rAsm.
This avoids a class of subtle bugs where code-gen mixed uses of
rScratch with calls to the assembler that could also make use of this
register.  (Tx64 code still uses rScratch/rAsm for now.)

While here, also moved the reserved XMM scratch registers from the
assembler to CodeGenerator, since the assembler doesn't really use
them.  (The uses of these registers in Tx64 were renamed back to
xmm0/xmm1 as they were until recently.)