Remove hoistGuardToLoad
This doesn't happen, and if it did it would be incorrect,
since it assumes a getTaken() implies a side exit which is no longer
true.  We may bring back something similar to this for LdClsCns but
it'll be easy to do it at that point (and a bunch of the instructions
in here aren't used as guards/checks anymore).